internal/src/main/kotlin/com/code42/jenkins/pipelinekt/internal/step/declarative/Tool.kt
package com.code42.jenkins.pipelinekt.internal.step.declarative

import com.code42.jenkins.pipelinekt.core.step.DeclarativeStep
import com.code42.jenkins.pipelinekt.core.step.SingletonStep
import com.code42.jenkins.pipelinekt.core.vars.Var
import com.code42.jenkins.pipelinekt.core.writer.GroovyWriter

data class Tool(val name: Var, val type: Var? = null) : DeclarativeStep, SingletonStep {
    override fun toGroovy(writer: GroovyWriter) {
        writer.writeln("tool name: ${name.toGroovy()}" + (type?.let { ", type: ${type.toGroovy()}" } ?: ""))
    }
}
